Simulation 2402h Meeting Rooms III -> Overflow: i32 :: i64
    Finished `test` profile [unoptimized + debuginfo] target(s) in 0.01s
     Running unittests src/lib.rs (target/debug/deps/simulation-ccd9f53bf6d0a41f)

running 1 test
* 2 [[0, 10], [1, 5], [2, 7], [3, 4]]
-> [(Reverse(0), 0), (Reverse(1), 0)]
-> O:   [(Reverse(10), Reverse(1), 2), (Reverse(11), Reverse(0), 2)]
-> F:   []
-> [(Reverse(2), 0), (Reverse(2), 1)]
:: 0
* 3 [[1, 20], [2, 10], [3, 5], [4, 9], [6, 8]]
-> [(Reverse(0), 0), (Reverse(1), 0), (Reverse(2), 0)]
-> O:   [(Reverse(10), Reverse(2), 2), (Reverse(20), Reverse(0), 1), (Reverse(12), Reverse(1), 2)]
-> F:   []
-> [(Reverse(2), 1), (Reverse(2), 2), (Reverse(1), 0)]
:: 1
test tests::test_2402 ... ok

test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 2 filtered out; finished in 0.00s

+++ simulation/src/lib.rs
@@ -44,7 +44,7 @@ impl Sol2402 {
         meetings.sort_unstable();
         for meeting in meetings {
             while let Some(&(Reverse(end), Reverse(room), count)) = ongoings.peek() {
-                if end <= meeting[0] {
+                if end <= meeting[0] as i64 {
                     ongoings.pop();
                     frees.push((Reverse(room), count));
                 } else {
@@ -53,10 +53,10 @@ impl Sol2402 {
             }
 
             if let Some((Reverse(room), count)) = frees.pop() {
-                ongoings.push((Reverse(meeting[1]), Reverse(room), count + 1));
+                ongoings.push((Reverse(meeting[1] as i64), Reverse(room), count + 1));
             } else if let Some((Reverse(end), Reverse(room), count)) = ongoings.pop() {
                 ongoings.push((
-                    Reverse(end + meeting[1] - meeting[0]),
+                    Reverse(end + (meeting[1] - meeting[0]) as i64),
                     Reverse(room),
                     count + 1,
                 ));
